--- src/components/class/board-contents/noticeboard/notice-write.css
@@ -1,9 +1,0 @@
-.buttons{
-    float: right;
-}
-.buttons:first-child{
-    margin: 0 16px 0 0;
-}
-.buttons + .buttons {
-    margin: 0 16px 0 0;
-}
